lfdd/libmem: fix unable to handle kernel paging request
Reading from high memory leads to unable to handle kernel paging request
error. It's because that the return value of phys_to_virt() is void*
and it leads to the if statement fails to compare the address if the
address is high(ex. 0xF0000000)

+++ lfdd/libmem.c
@@ -184,7 +184,7 @@ void lfdd_mem_read_256byte( struct lfdd_mem_t *pmem ) {
 
     // Check the range of physical address
     if( ((0xffffffff - pmem->addr) <= LFDD_MASSBUF_SIZE) 
-        || ((pmem->addr + LFDD_MASSBUF_SIZE) >= virt_to_phys( high_memory )) ) {
+        || ((pmem->addr + LFDD_MASSBUF_SIZE) >= (unsigned int)virt_to_phys( high_memory )) ) {
 
         virtmem = ioremap( pmem->addr, LFDD_MASSBUF_SIZE );
